refactor(nano): Split runner in multiples files
